Refactor: DB - Moved Interfaces
Moved Interfaces to types/models.ts
